requirements.txt
ipykernel
numpy 
pandas
scikit-learn
matplotlib
seaborn
fastapi
uvicorn
sqlalchemy
pydantic
langchain 
chromadb 
sentence-transformers 
PyPDF2 
huggingface-hub
python-dotenv
joblib
langchain-community
langchain-text-splitters
langchain-core
langchain-huggingface